+++ docs/changelog.rst
@@ -18,6 +18,8 @@ Chula v0.12.0 (dev)
 *Still under development*
 
 - Add support for using the Gunicorn wsgi provider.
+- Add support for "restfull" urls.  This feature is only available
+  when using the regex mapper, see: :ref:`configuration`
 - Don't include a stacktrace for 404 requests served by
   :meth:`chula.www.controller.error.Error._crappy_static_server`.